server/requirements.txt
django
django-cors-headers
numpy
scikit-learn
nltk
bson
pymongo
tqdm
djangorestframework_simplejwt